Add a real epoch DSL for RPM and DEB builds
This replaces the dirty hack that was there before and actually makes
things work for real with RPM packages as explained below.

The epoch shouldn't be included as part of the version string for RPMs
but as a separate field in the package metadata (contrary to what's
happening for debs). I updated our fork by actually (re)adding an epoch
DSL for RPM and DEB builds which is definitely cleaner than what we were
doing before. Will make a PR for that against `chef/omnibus` when I get
the chance, it might definitely be a relevant DSL to other Omnibus users.