Fixed the RPM build when the epoch DSL is unset
The build was broken because only `:` was inserted into the RPM package
name when no epoch was set which wasn't caught by the `\d+:` RegEx. No
when no epoch is set, no epoch is passed to the `rpmbuild` command and
the RPM doesn't have any extra colon. When it is defined, the epoch is
removed by the regex as it was before :)