[Windows] Allow extra package files to be shipped
We don't have an `extra_package_file` directive for Windows builds
(there's one but it doesn't do anything) and as a matter of fact, it
wouldn't make much sense since the directories in whcih we'd like to put
these files are often User/System dependant or user configurable.

We added a routine to add such files under a directory defined in the
`source.wxs` file and referenced by it's id. To achieve that, the only
thing that needs to be done is to copy the extra files under
`<main_software_source_dir>/EXTRA_PACKAGE_FILES/DIRECTORY_ID`.

I'll try to figure out a cleaner way and have it packaged in the next
Omnibus release but for now this does the trick pretty well :)