[version] Replaced + by period to comply with S3
S3 doesn't accept `+` characters and they get replaced by a space when
queried in a URL. It somehow occured yesterday, no idea why or how.

Even though it's not semver compliant, I replaced that `+` with a `.`,
which plays nicely with S3.

Note that it's only impacting nightly builds for now (so it can't have
any sort of influence on our customers) but if we want to add semver
compliant `beta`s, we might have to think about it a bit more.

Also note that according to a couple of tickets on the web, apt version
`0.9.9.1~ubuntu1` but staging is impacted by the bug even though we're
using a higher version there... so I have some doubts.

Here's the ticket :
https://bugs.launchpad.net/ubuntu/+source/apt/+bug/1003633

It also suggests an interesting workaround but it would require us to
drop `deb-s3` to upload our packages :/