Revert changes to versioning DSL (dd_agent_format, s3_compliant_semver)
We go back to using the previous versioning scheme, because
we cannot rely only on the tags of the current branch to determine
the version of the agent for nightlies.

For instance when a bugfix release 5.5.1 has been released, the `5.5.1`
tag is only on the `5.5.x` branch, so nightlies will still be versioned
`5.5.0.git...` and have a lower version than the latest stable.

With this revert we go back to using the env var (and hence the version
in `config.py` of dd-agent) to determine the version of nightlies.

This reverts commits 1b9753b5d43d0d1c70c77fc8acaac50bc6a6bf7e
and 7875a77ae5e5daebd9bfd505b268c0a522106278, reversing
changes made to 63672ffde2eb51ade456bd3acd4d1136070ed618.