[version] add dd_agent_format for dd-agent
We cannot rely only on the tags of the current branch to determine
the version of the agent for nightlies.

For instance when a bugfix release 5.5.1 has been released, the `5.5.1`
tag is only on the `5.5.x` branch, so nightlies will still be versioned
`5.5.0.git...` and have a lower version than the latest stable.

See 8c354c06e5f3529435fc6eaf1ed4ff8a0c947d6b